feat(tui): live hit resolution for scrollable timeline content
Timeline tool hits are now resolved at event time from content-space row
ownership plus the live viewport offset, instead of absolute screen rects
baked into the per-frame hit map. Scroll batches can no longer make a click
or hover stale, and hits use stable interned tool ids rather than component
slots, so projection rebuilds never retarget a click.

- Timeline interns stable tool hit ids and bumps a layout epoch on every
  geometry-changing mutation; pure scroll never invalidates the retained plan.
- TimelineRenderPlan carries row_owners/content-space hits; the pointer path
  refreshes the plan only on epoch mismatch and reads top_offset live.
- Hover re-derives from the last pointer position after viewport changes.
- PRD/design: packages/tui/docs/{features,design}/pointer-hit-validity.md.
- Regression tests cover wheel-batch+click, expand+click, streaming append,
  identity across rebuilds, hover re-derivation, and banner/gap rows.

+++ packages/tui/src/app/command.rs
@@ -83,7 +83,8 @@ pub enum TimelineAction {
     ScrollUp(usize),
     ScrollDown(usize),
     JumpLatest,
-    ToggleTool(usize),
+    /// Toggle one tool block by its stable interned hit id.
+    ToggleTool(u64),
 }
 
 #[derive(Debug)]
